TemplateService/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
minio==7.2.0
python-dotenv==1.0.0
pydantic>=2.10.3
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
python-pptx==0.6.23
Pillow==10.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
requests>=2.31.0
tabulate>=0.9.0
